#26: Move registration screen to registration_form package
