+++ android/app/src/main/java/ru/gortea/petter/MainActivity.kt
@@ -13,7 +13,7 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.updatePadding
 import ru.gortea.petter.arch.android.compose.LocalApplicationContext
-import ru.gortea.petter.auth.registration.ui.RegistrationScreen
+import ru.gortea.petter.auth.registration.registration_form.ui.RegistrationScreen
 import ru.gortea.petter.theme.PetterAppTheme
 
 class MainActivity : ComponentActivity() {
@@ -23,6 +23,7 @@ class MainActivity : ComponentActivity() {
 
         setupInsets()
     }
+
     @Composable
     private fun Content() {
         CompositionLocalProvider(LocalApplicationContext provides applicationContext) {
@@ -38,7 +39,7 @@ class MainActivity : ComponentActivity() {
     }
 
     private fun setupInsets() {
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(android.R.id.content)){ view, insets ->
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(android.R.id.content)) { view, insets ->
             val bottom = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
             view.updatePadding(bottom = bottom)
             insets
