+++ android/modules/features/auth-data/src/main/java/ru/gortea/petter/auth/data/RegistrationRepository.kt
@@ -5,6 +5,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import ru.gortea.petter.auth.data.api.AuthApi
 import ru.gortea.petter.auth.data.model.RegistrationModel
+import ru.gortea.petter.auth.data.model.RegistrationSuccessModel
 import ru.gortea.petter.data.SourceRepository
 import ru.gortea.petter.data.model.DataState
 
@@ -22,7 +23,7 @@ class RegistrationRepository(
     fun retryCreateAccount(args: RegistrationModel) {
         registrationSource.invalidate(args)
     }
-    suspend fun createAccount(args: RegistrationModel): Flow<DataState<String>> {
+    suspend fun createAccount(args: RegistrationModel): Flow<DataState<RegistrationSuccessModel>> {
         return registrationSource.get(args)
     }
 }
