+++ android/modules/features/auth-data/src/main/java/ru/gortea/petter/auth/data/api/AuthApi.kt
@@ -2,10 +2,15 @@ package ru.gortea.petter.auth.data.api
 
 import retrofit2.http.Body
 import retrofit2.http.POST
+import ru.gortea.petter.auth.data.model.RegistrationConfirmModel
 import ru.gortea.petter.auth.data.model.RegistrationModel
+import ru.gortea.petter.auth.data.model.RegistrationSuccessModel
 
 interface AuthApi {
 
     @POST("api/registration")
-    suspend fun registration(@Body model: RegistrationModel): String
+    suspend fun registration(@Body model: RegistrationModel): RegistrationSuccessModel
+
+    @POST("api/registration/confirm")
+    suspend fun registrationConfirm(@Body model: RegistrationConfirmModel)
 }
