+++ android/modules/features/auth-data/src/main/java/ru/gortea/petter/auth/data/model/RegistrationSuccessModel.kt
@@ -1,0 +1,8 @@
+package ru.gortea.petter.auth.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RegistrationSuccessModel(
+    val id: String
+)
