--- android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/presentation/actors/RegistrationCreateAccountActor.kt
@@ -1,19 +1,0 @@
-package ru.gortea.petter.auth.registration.presentation.actors
-
-import kotlinx.coroutines.flow.*
-import ru.gortea.petter.arch.Actor
-import ru.gortea.petter.auth.data.RegistrationRepository
-import ru.gortea.petter.auth.registration.presentation.RegistrationCommand
-import ru.gortea.petter.auth.registration.presentation.RegistrationEvent
-
-internal class RegistrationCreateAccountActor(
-    private val repository: RegistrationRepository
-) : Actor<RegistrationCommand, RegistrationEvent> {
-
-    override fun process(commands: Flow<RegistrationCommand>): Flow<RegistrationEvent> {
-        return commands
-            .filterIsInstance<RegistrationCommand.CreateAccount>()
-            .flatMapMerge { repository.createAccount(it.model) }
-            .map { RegistrationEvent.AccountCreateProcess(it) }
-    }
-}
