+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/presentation/RegistrationCommand.kt
@@ -1,7 +1,7 @@
-package ru.gortea.petter.auth.registration.presentation
+package ru.gortea.petter.auth.registration.registration_form.presentation
 
 import ru.gortea.petter.auth.data.model.RegistrationModel
-import ru.gortea.petter.auth.registration.presentation.state.RegistrationState
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.RegistrationState
 
 internal sealed interface RegistrationCommand {
     class Validate(val state: RegistrationState) : RegistrationCommand
