+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/presentation/RegistrationStore.kt
@@ -1,13 +1,13 @@
-package ru.gortea.petter.auth.registration.presentation
+package ru.gortea.petter.auth.registration.registration_form.presentation
 
 import ru.gortea.petter.arch.store.MviStore
 import ru.gortea.petter.arch.store.factory.TeaStore
 import ru.gortea.petter.auth.registration.di.RegistrationComponent
-import ru.gortea.petter.auth.registration.presentation.actors.RegistrationCreateAccountActor
-import ru.gortea.petter.auth.registration.presentation.actors.RegistrationRetryCreateAccountActor
-import ru.gortea.petter.auth.registration.presentation.actors.RegistrationValidateActor
-import ru.gortea.petter.auth.registration.presentation.state.RegistrationState
-import ru.gortea.petter.auth.registration.presentation.validation.RegistrationValidatorComposite
+import ru.gortea.petter.auth.registration.registration_form.presentation.actors.RegistrationCreateAccountActor
+import ru.gortea.petter.auth.registration.registration_form.presentation.actors.RegistrationRetryCreateAccountActor
+import ru.gortea.petter.auth.registration.registration_form.presentation.actors.RegistrationValidateActor
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.RegistrationState
+import ru.gortea.petter.auth.registration.registration_form.presentation.validation.RegistrationValidatorComposite
 
 internal typealias RegistrationStore = MviStore<RegistrationState, RegistrationEvent, Nothing>
