+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/presentation/actors/RegistrationValidateActor.kt
@@ -1,14 +1,14 @@
-package ru.gortea.petter.auth.registration.presentation.actors
+package ru.gortea.petter.auth.registration.registration_form.presentation.actors
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.filterIsInstance
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapLatest
 import ru.gortea.petter.arch.Actor
-import ru.gortea.petter.auth.registration.presentation.RegistrationCommand
-import ru.gortea.petter.auth.registration.presentation.RegistrationEvent
-import ru.gortea.petter.auth.registration.presentation.state.RegistrationState
-import ru.gortea.petter.auth.registration.presentation.validation.reason.RegistrationFailedReason
+import ru.gortea.petter.auth.registration.registration_form.presentation.RegistrationCommand
+import ru.gortea.petter.auth.registration.registration_form.presentation.RegistrationEvent
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.RegistrationState
+import ru.gortea.petter.auth.registration.registration_form.presentation.validation.reason.RegistrationFailedReason
 import ru.gortea.petter.auth.validation.ValidatorComposite
 
 internal class RegistrationValidateActor(
