+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/presentation/state/RegistrationState.kt
@@ -1,5 +1,6 @@
-package ru.gortea.petter.auth.registration.presentation.state
+package ru.gortea.petter.auth.registration.registration_form.presentation.state
 
+import ru.gortea.petter.auth.data.model.RegistrationSuccessModel
 import ru.gortea.petter.data.model.DataState
 
 internal data class RegistrationState(
@@ -7,5 +8,5 @@ internal data class RegistrationState(
     val username: RegistrationFieldState = RegistrationFieldState("jmax"),
     val password: RegistrationFieldState = RegistrationFieldState("1234").hide(),
     val passwordConfirm: RegistrationFieldState = RegistrationFieldState("1234").hide(),
-    val registrationStatus: DataState<String> = DataState.Empty
+    val registrationStatus: DataState<RegistrationSuccessModel> = DataState.Empty
 )
