+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/presentation/validation/validators/PasswordValidator.kt
@@ -1,7 +1,7 @@
-package ru.gortea.petter.auth.registration.presentation.validation.validators
+package ru.gortea.petter.auth.registration.registration_form.presentation.validation.validators
 
-import ru.gortea.petter.auth.registration.presentation.state.RegistrationState
-import ru.gortea.petter.auth.registration.presentation.state.isNotBlank
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.RegistrationState
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.isNotBlank
 import ru.gortea.petter.auth.validation.Validator
 
 internal class PasswordValidator : Validator<RegistrationState> {
