+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/ui/RegistrationScreen.kt
@@ -1,4 +1,4 @@
-package ru.gortea.petter.auth.registration.ui
+package ru.gortea.petter.auth.registration.registration_form.ui
 
 import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.layout.*
@@ -21,11 +21,11 @@ import ru.gortea.petter.arch.android.compose.storeHolder
 import ru.gortea.petter.arch.android.store.getValue
 import ru.gortea.petter.auth.R
 import ru.gortea.petter.auth.registration.di.RegistrationComponent
-import ru.gortea.petter.auth.registration.presentation.RegistrationStore
-import ru.gortea.petter.auth.registration.presentation.RegistrationUiEvent.*
-import ru.gortea.petter.auth.registration.presentation.createRegistrationStore
-import ru.gortea.petter.auth.registration.ui.mapper.RegistrationUiStateMapper
-import ru.gortea.petter.auth.registration.ui.state.RegistrationUiState
+import ru.gortea.petter.auth.registration.registration_form.presentation.RegistrationStore
+import ru.gortea.petter.auth.registration.registration_form.presentation.RegistrationUiEvent.*
+import ru.gortea.petter.auth.registration.registration_form.presentation.createRegistrationStore
+import ru.gortea.petter.auth.registration.registration_form.ui.mapper.RegistrationUiStateMapper
+import ru.gortea.petter.auth.registration.registration_form.ui.state.RegistrationUiState
 import ru.gortea.petter.theme.PetterAppTheme
 import ru.gortea.petter.theme.appHeader
 import ru.gortea.petter.ui_kit.TextField
@@ -106,7 +106,8 @@ private fun RegistrationScreenContent(
     modifier: Modifier = Modifier
 ) {
     Column(
-        modifier = modifier.fillMaxSize()
+        modifier = modifier
+            .fillMaxSize()
             .verticalScroll(rememberScrollState()),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
