+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/ui/mapper/RegistrationUiStateMapper.kt
@@ -1,13 +1,13 @@
-package ru.gortea.petter.auth.registration.ui.mapper
+package ru.gortea.petter.auth.registration.registration_form.ui.mapper
 
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import ru.gortea.petter.arch.UiStateMapper
-import ru.gortea.petter.auth.registration.presentation.state.RegistrationFieldState
-import ru.gortea.petter.auth.registration.presentation.state.RegistrationState
-import ru.gortea.petter.auth.registration.ui.state.ButtonState
-import ru.gortea.petter.auth.registration.ui.state.RegistrationUiState
-import ru.gortea.petter.auth.registration.ui.state.TextFieldState
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.RegistrationFieldState
+import ru.gortea.petter.auth.registration.registration_form.presentation.state.RegistrationState
+import ru.gortea.petter.auth.registration.registration_form.ui.state.ButtonState
+import ru.gortea.petter.auth.registration.registration_form.ui.state.RegistrationUiState
+import ru.gortea.petter.auth.registration.registration_form.ui.state.TextFieldState
 import ru.gortea.petter.data.model.DataState
 
 internal class RegistrationUiStateMapper : UiStateMapper<RegistrationState, RegistrationUiState> {
