+++ android/modules/features/auth/src/main/java/ru/gortea/petter/auth/registration/registration_form/ui/state/ButtonState.kt
@@ -1,4 +1,4 @@
-package ru.gortea.petter.auth.registration.ui.state
+package ru.gortea.petter.auth.registration.registration_form.ui.state
 
 internal data class ButtonState(
     val isLoading: Boolean = false
